Add files via upload
This was an exciting Project of data cleaning and data visualization even though I mostly use Tableau for my visualizations but I created a dashboard in the Excel project and added some splicers to it. Let me begin by saying I was excited working on this project because I got the chance to work on 534 records and firstly, I had to split the columns and the freeze the columns and unfreeze them. Then I had to locate or highlight some duplicates using the Conditional Formatting highlighting cell rule and then Duplicate values. I removed or deleted some duplicates also. It was fun because I wanted to clean my data and make use of it. I had to sort out some data in a range, multiple fields, create subtotals for the rows, filter a create some excel table and create a dashboard just to say a few.